consolidated redux reducers in a single store
