+++ app/home/autoComplete/page.tsx
@@ -1,8 +1,7 @@
 'use client';
-import { Provider } from 'react-redux';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { store } from '@/app/stores';
+import { StoreProvider } from '@/app/stores/provider';
 import AutoComplete from './AutoComplete';
 
 // Create a client
@@ -17,12 +16,12 @@ const queryClient = new QueryClient({
 
 function App() {
 	return (
-		<Provider store={store}>
+		<StoreProvider>
 			<QueryClientProvider client={queryClient}>
 				<AutoComplete />
 				<ReactQueryDevtools initialIsOpen={false} />
 			</QueryClientProvider>
-		</Provider>
+		</StoreProvider>
 	);
 }
