+++ app/hooks/index.ts
@@ -1,10 +1,10 @@
-export { useRenderCounter } from '@/app/hooks/useRenderCounter';
-export { useSearchParams } from '@/app/hooks/useSearchParams';
-export { useEventListener } from '@/app/hooks/useEventListener';
-export { useTimer } from '@/app/hooks/useTimer';
-export { useClickOutside } from '@/app/hooks/useClickOutside';
-export { useToggle } from '@/app/hooks/useToggle';
-export { useDebouncedCallback } from '@/app/hooks/useDebouncedCallback';
-export { useDebounced } from '@/app/hooks/useDebounced';
-export { useCompanySearch } from '@/app/hooks/useCompanySearch';
-export { useAppDispatch, useAppSelector } from '@/app/hooks/redux';
+export { useRenderCounter } from './useRenderCounter';
+export { useSearchParams } from './useSearchParams';
+export { useEventListener } from './useEventListener';
+export { useTimer } from './useTimer';
+export { useClickOutside } from './useClickOutside';
+export { useToggle } from './useToggle';
+export { useDebouncedCallback } from './useDebouncedCallback';
+export { useDebounced } from './useDebounced';
+export { useCompanySearch } from './useCompanySearch';
+export { useAppDispatch, useAppSelector } from './redux';
